backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-plugin-utils==0.2.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.0.0
wrapt==1.11.2
